Updated the example site_configs to have slightly less random names.
Changed the dev_env variable to a more verbose "development_environment" and removed the resetting of the postgres user password. This *only* happens now if the password is provided and development_environment is set to true. In a production env we don't need the postgres user to have a password - its simply a useful thing for development but it should still be optional.